Overhauled the program options. Improved the performance of the concurrency::parallel_for utility function. Improved the performance of the RayTracer by removing the container of projections for use in accumulation of positive values. The accumulation is now simply performed without any intermediate steps. Switched the underlying vector_type from float to double due to no loss in performance from a greater floating-point precision.

+++ source/raytracer/include/raytracer/alias.hpp
@@ -26,5 +26,5 @@
 namespace raytracer
 {
 	// Type for use as underlying render values.
-	using vector_type = float;
+	using vector_type = double;
 }
